# docs/swagger.yaml
basePath: /
definitions:
  domain.Book:
    properties:
      author:
        type: string
      id:
        type: integer
      publish_date:
        type: string
      rating:
        type: string
      title:
        type: string
    type: object
  domain.UpdateBookInput:
    properties:
      author:
        type: string
      publish_date:
        type: string
      rating:
        type: integer
      title:
        type: string
    type: object
  handler.errorResponse:
    properties:
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: CRUD Server for Books Application
  title: CRUD App API
  version: "1.0"
paths:
  /book:
    delete:
      consumes:
      - application/json
      description: create book
      produces:
      - application/json
      responses:
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Delete book
      tags:
      - Books
    get:
      consumes:
      - application/json
      description: Get book by ID
      produces:
      - application/json
      responses:
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Get book
      tags:
      - Books
    post:
      consumes:
      - application/json
      description: create book
      operationId: createBook
      parameters:
      - description: Book info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.Book'
      produces:
      - application/json
      responses:
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Create book
      tags:
      - Books
    put:
      consumes:
      - application/json
      description: create book
      parameters:
      - description: Book info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateBookInput'
      produces:
      - application/json
      responses:
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Update book
      tags:
      - Books
  /books:
    get:
      consumes:
      - application/json
      description: create book
      produces:
      - application/json
      responses:
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: GetAll
      tags:
      - Books
swagger: "2.0"